+++ server/src/main/java/com/lankacapital/server/repositories/InstallmentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InstallmentRepository extends JpaRepository<Installment, Double> {
+public interface InstallmentRepository extends JpaRepository<Installment, Integer> {
     Installment findByValue(Integer value);
 }
